[MIOpen] Replace legacy cba_infer gtest with migrated implementation.
Keep the original cba_infer.cpp filename and remove cba_inference.cpp to avoid duplicate gtest symbols without changing test/gtest/CMakeLists.txt.
